Added Support for Basic Micro DaVinci/Orion
WIP - This is a Work in progress.  Playing with a Basic Micro (Orion
Robotics) DaVinci board which is an Arduino 328p based board with an
Orion Shield.  This shield has 32 IO pins, 24 of them are controlled by
the shield which is more or less an ARC32...  I am using libraries and
code from Basic Micro.