PhantomX V2 - Works on mine now :)
I did a minimalist change here to make the V2 work.  The remaining issue
that was not addressed in the first pass was the Tibia servos move in
opposite direction on V2 versus V1 so needed to change AX12 driver to do
this conditionally.  Note: I did this a different way in the Quad
branch...