RC Control input update.
The previous version did not work with Arbotix (Atmega 644p) processors
as the Pin Change Interrupt code did not work properly on mapping back
from PCINT0,1,2,3 interrupts to which IOPort is associated with it.
Checked update with Arbotix, Lynxmotion Botboarduino, and Arduino
Leonardo.  Note: comments were updated in function to show which pins
are supported (I think).